Moved tests to correct namespaces
